backend/pyproject.toml
[project]
name = "canlogger-backend"
version = "0.1.0"
description = "Multi-scope CAN waveform logger backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "orjson>=3.10.0",
    "pyyaml>=6.0.2",
    "numpy>=2.1.0",
    "psycopg[binary]>=3.2.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]
